fix: Refactor App - split on more parts
