+++ src/components/App/index.tsx
@@ -1,10 +1,8 @@
 import type {NavigationData, PageContent} from '@gravity-ui/page-constructor';
 import type {ReactElement} from 'react';
 import type {Props as HeaderControlsProps} from '../HeaderControls';
-import type {Settings} from './useSettings';
 
-import React, {useCallback, useEffect, useMemo} from 'react';
-import {PageConstructor, PageConstructorProvider} from '@gravity-ui/page-constructor';
+import React, {useEffect} from 'react';
 import {ThemeProvider} from '@gravity-ui/uikit';
 import {
     ConsentPopup,
@@ -18,16 +16,12 @@ import {
 } from '@diplodoc/components';
 import '@diplodoc/transform/dist/js/yfm';
 
-import {HeaderControls} from '../HeaderControls';
 import {getDirection, updateRootClassName, updateThemeClassName} from '../../utils';
-import {ConstructorPage} from '../ConstructorPage';
-import {useContent} from '../ConstructorPage/useContent';
-import {useNavigation} from '../ConstructorPage/useNavigation';
 import {LangProvider} from '../../hooks/useLang';
 import '../../interceptors/leading-page-links';
 import '../../interceptors/fast-class-applier';
 
-import {Page} from './Page';
+import {LegacyNavPage, RichNavPage} from './Page';
 import {Runtime} from './Runtime';
 import {useLangs} from './useLangs';
 import {useSettings} from './useSettings';
@@ -61,81 +55,6 @@ export type DocInnerProps<Data extends PageData = PageData> = {
 } & AppProps;
 
 export type {DocLeadingPageData, DocPageData};
-
-type PageProps<T extends {} = {}> = {
-    data: DocBasePageData<T> & PageData;
-    props: {
-        router: AppProps['router'];
-    } & Settings;
-    controls: HeaderControlsProps;
-};
-
-const LegacyNavPage = ({data, props, controls}: PageProps) => {
-    const {theme} = props;
-
-    const CustomPage = useCallback(
-        () => <ConstructorPage {...(data as DocContentPageData).data} />,
-        [data],
-    );
-    const content = useContent(data as DocContentPageData, CustomPage);
-
-    const custom = useMemo(
-        () => ({
-            blocks: content.custom,
-        }),
-        [content],
-    );
-
-    return (
-        <Page data={data} headerHeight={0} {...props} {...controls}>
-            <PageConstructorProvider theme={theme}>
-                <PageConstructor custom={custom} content={content.layout} />
-            </PageConstructorProvider>
-        </Page>
-    );
-};
-
-const empty = {};
-
-const RichNavPage = ({data, props, controls}: PageProps<WithNavigation>) => {
-    const {theme, fullScreen} = props;
-
-    const CustomControls = useCallback(() => <HeaderControls {...controls} />, [controls]);
-    const navigation = useNavigation(data, CustomControls);
-
-    const CustomPage = useCallback(
-        () => (
-            <Page
-                data={data}
-                headerHeight={fullScreen ? 0 : 64}
-                {...props}
-                {...(navigation.withControls ? empty : controls)}
-            >
-                <ConstructorPage {...(data as DocContentPageData).data} />
-            </Page>
-        ),
-        [data, navigation, props, controls, fullScreen],
-    );
-    const content = useContent(data as DocContentPageData, CustomPage);
-
-    const custom = useMemo(
-        () => ({
-            navigation: navigation.custom,
-            blocks: content.custom,
-        }),
-        [navigation, content],
-    );
-
-    return (
-        <PageConstructorProvider theme={theme}>
-            <PageConstructor
-                custom={custom}
-                content={content.layout}
-                navigation={fullScreen ? undefined : navigation.layout}
-            />
-        </PageConstructorProvider>
-    );
-};
 
 function hasNavigation(
     data: DocBasePageData<Partial<WithNavigation>>,
@@ -179,8 +98,8 @@ export function App(props: DocInnerProps): ReactElement {
 
     return (
         <div className="App">
-            <LangProvider value={lang}>
-                <ThemeProvider theme={theme} direction={direction}>
+            <ThemeProvider theme={theme} direction={direction}>
+                <LangProvider value={lang}>
                     {hasNavigation(data) ? (
                         <RichNavPage data={data} props={page} controls={controls} />
                     ) : (
@@ -193,9 +112,9 @@ export function App(props: DocInnerProps): ReactElement {
                             consentMode={analytics?.gtm?.mode}
                         />
                     )}
-                </ThemeProvider>
-            </LangProvider>
-            <Runtime theme={theme} />
+                    <Runtime />
+                </LangProvider>
+            </ThemeProvider>
         </div>
     );
 }
